requirements.txt
langchain
langchain_community
langchain_Groq
wikipedia
streamlit
groq
python-dotenv
